Crate rustc_lint

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Lints in the Rust compiler.

This currently only contains the definitions and implementations of most of the lints that rustc supports directly, it does not contain the infrastructure for defining/registering lints. That is available in rustc::lint and rustc_plugin respectively.

Note

This API is completely unstable and subject to change.

Functions

register_builtins 	 [
Experimental
]
Tell the LintStore about all the built-in lints (the ones defined in this crate and the ones defined in rustc::lint::builtin).