rustc_lint

Module dangling

Structs

DanglingPointerLocalContext 🔒

DanglingPointerReturnSearcher 🔒

DanglingPointerSearcher 🔒

This produces a dangling pointer:

DanglingPointers 🔒

FIXME: false negatives (i.e. the lint is not emitted when it should be)

Statics

DANGLING_POINTERS_FROM_LOCALS

The dangling_pointers_from_locals lint detects getting a pointer to data of a local that will be dropped at the end of the function.

DANGLING_POINTERS_FROM_TEMPORARIES

The dangling_pointers_from_temporaries lint detects getting a pointer to data of a temporary that will immediately get dropped.

Functions

is_temporary_rvalue 🔒

lint_addr_of_local 🔒

Look for &<path_to_local_in_same_body> pattern and emit lint for it

lint_expr 🔒

owns_allocation 🔒