rustc_lint
Module drop_forget_useless

Structs

DropForgetUseless

Statics

DROPPING_COPY_TYPES

The dropping_copy_types lint checks for calls to std::mem::drop with a value that derives the Copy trait.

DROPPING_REFERENCES

The dropping_references lint checks for calls to std::mem::drop with a reference instead of an owned value.

FORGETTING_COPY_TYPES

The forgetting_copy_types lint checks for calls to std::mem::forget with a value that derives the Copy trait.

FORGETTING_REFERENCES

The forgetting_references lint checks for calls to std::mem::forget with a reference instead of an owned value.

UNDROPPED_MANUALLY_DROPS

The undropped_manually_drops lint check for calls to std::mem::drop with a value of std::mem::ManuallyDrop which doesn’t drop.

Functions

is_single_call_in_arm 🔒