rustc_lint

Module if_let_rescope

Structs

AltHead 🔒

ClosingBrackets 🔒

ConsequentRewrite 🔒

DestructorLabel 🔒

FindSignificantDropper 🔒

IfLetRescope 🔒

Lint for potential change in program semantics of if lets

IfLetRescopeLint 🔒

IfLetRescopeRewrite 🔒

Enums

SingleArmMatchBegin 🔒

Statics

IF_LET_RESCOPE

The if_let_rescope lint detects cases where a temporary value with significant drop is generated on the right hand side of if let and suggests a rewrite into match when possible.

Functions

expr_parent_is_else 🔒

expr_parent_is_stmt 🔒

match_head_needs_bracket 🔒