rustc_lint

Module impl_trait_overcaptures

Structs

FunctionalVariances 🔒

Computes the variances of regions that appear in the type, but considering late-bound regions too, which don’t have their variance computed usually.

ImplTraitOvercaptures

Lint for opaque types that will begin capturing in-scope but unmentioned lifetimes in edition 2024.

ImplTraitOvercapturesLint 🔒

ImplTraitRedundantCapturesLint 🔒

VisitOpaqueTypes 🔒

Enums

ParamKind 🔒

Statics

IMPL_TRAIT_OVERCAPTURES

The impl_trait_overcaptures lint warns against cases where lifetime capture behavior will differ in edition 2024.

IMPL_TRAIT_REDUNDANT_CAPTURES

The impl_trait_redundant_captures lint warns against cases where use of the precise capturing use<...> syntax is not needed.

Functions

check_fn 🔒

extract_def_id_from_arg 🔒

unify 🔒

What is the variance that satisfies the two variances?