Module rustc_lint::internal

Some lints that are only useful in the compiler or crates that use compiler internals, such as Clippy.

Structs

BadOptAccess

DefaultHashTypes

Diagnostics

ExistingDocKeyword

LintPassImpl

QueryStability

TyTyKind

Statics

BAD_OPT_ACCESS

DEFAULT_HASH_TYPES

DIAGNOSTIC_OUTSIDE_OF_IMPL

EXISTING_DOC_KEYWORD

LINT_PASS_IMPL_WITHOUT_MACRO

POTENTIAL_QUERY_INSTABILITY

UNTRANSLATABLE_DIAGNOSTIC

USAGE_OF_QUALIFIED_TY

USAGE_OF_TY_TYKIND

Functions

gen_args 🔒

is_doc_keyword 🔒

is_ty_or_ty_ctxt 🔒

lint_ty_kind_usage 🔒

typeck_results_of_method_fn 🔒

Helper function for lints that check for expressions with calls and use typeck results to get the DefId and SubstsRef of the function.