rustc_lint
Module let_underscore

Structs

LetUnderscore

Constants

SYNC_GUARD_SYMBOLS 🔒

Statics

LET_UNDERSCORE_DROP

The let_underscore_drop lint checks for statements which don’t bind an expression which has a non-trivial Drop implementation to anything, causing the expression to be dropped immediately instead of at end of scope.

LET_UNDERSCORE_LOCK

The let_underscore_lock lint checks for statements which don’t bind a mutex to anything, causing the lock to be released immediately instead of at end of scope, which is typically incorrect.