Module rustc_lint::levels

Structs

BuilderPush ğŸ”’

LintLevelQueryMap ğŸ”’

LintLevelSets ğŸ”’

Collection of lint levels for the whole crate. This is used by AST-based lints, which do not wait until we have built HIR to be emitted.

LintLevelsBuilder

LintSet ğŸ”’

Specifications found at this position in the stack. This map only represents the lints found for one set of attributes (like shallow_lint_levels_on does).

LintStackIndex ğŸ”’

QueryMapExpectationsWrapper ğŸ”’

TopDown

Constants

COMMAND_LINE ğŸ”’

Traits

LintLevelsProvider

Functions

lint_expectations ğŸ”’

provide ğŸ”’

shallow_lint_levels_on ğŸ”’