Module rustc_lint::lint::builtin

Unstable

Some lints that are built in to the compiler.

These are the built-in lints that are emitted direct in the main compiler code, rather than using their own custom pass. Those lints are all available in rustc_lint::builtin.

Structs

HardwiredLints 	 [Unstable]
Does nothing as a lint pass, but registers some Lints which are used by other parts of the compiler.

Statics

DEAD_CODE 	 [Unstable]

FAT_PTR_TRANSMUTES 	 [Unstable]

STABLE_FEATURES 	 [Unstable]

TRIVIAL_CASTS 	 [Unstable]

TRIVIAL_NUMERIC_CASTS 	 [Unstable]

UNKNOWN_CRATE_TYPES 	 [Unstable]

UNKNOWN_LINTS 	 [Unstable]

UNREACHABLE_CODE 	 [Unstable]

UNUSED_ASSIGNMENTS 	 [Unstable]

UNUSED_EXTERN_CRATES 	 [Unstable]

UNUSED_FEATURES 	 [Unstable]

UNUSED_IMPORTS 	 [Unstable]

UNUSED_QUALIFICATIONS 	 [Unstable]

UNUSED_VARIABLES 	 [Unstable]

VARIANT_SIZE_DIFFERENCES 	 [Unstable]

WARNINGS 	 [Unstable]