Struct rustc_lint::lint::builtin::HardwiredLints

pub struct HardwiredLints;

Unstable (rustc_private)

Does nothing as a lint pass, but registers some Lints which are used by other parts of the compiler.