Module rustc_lint::metadata::csearch

Unstable

Structs

MethodInfo 	 [Unstable]

Enums

FoundAst 	 [Unstable]

Functions

each_child_of_item 	 [Unstable]
Iterates over each child of the given item.

each_implementation_for_trait 	 [Unstable]

each_inherent_implementation_for_type 	 [Unstable]

each_lang_item 	 [Unstable]
Iterates over all the language items in the given crate.

each_top_level_item_of_crate 	 [Unstable]
Iterates over each top-level crate item.

get_associated_consts 	 [Unstable]

get_custom_coerce_unsized_kind 	 [Unstable]

get_dylib_dependency_formats 	 [Unstable]

get_enum_variants 	 [Unstable]

get_field_type 	 [Unstable]

get_impl_items 	 [Unstable]
Returns information about the given implementation.

get_impl_or_trait_item 	 [Unstable]

get_impl_polarity 	 [Unstable]

get_impl_trait 	 [Unstable]

get_item_attrs 	 [Unstable]

get_item_path 	 [Unstable]

get_item_variances 	 [Unstable]

get_method_arg_names 	 [Unstable]

get_methods_if_impl 	 [Unstable]

get_missing_lang_items 	 [Unstable]

get_native_libraries 	 [Unstable]

get_predicates 	 [Unstable]

get_provided_trait_methods 	 [Unstable]

get_reachable_extern_fns 	 [Unstable]

get_repr_attrs 	 [Unstable]

get_stability 	 [Unstable]

get_struct_field_attrs 	 [Unstable]

get_struct_fields 	 [Unstable]

get_super_predicates 	 [Unstable]

get_symbol 	 [Unstable]

get_trait_def 	 [Unstable]

get_trait_item_def_ids 	 [Unstable]

get_trait_name 	 [Unstable]

get_trait_of_item 	 [Unstable]
If the given def ID describes an item belonging to a trait (either a default method or an implementation of a trait method), returns the ID of the trait that the method belongs to. Otherwise, returns None.

get_tuple_struct_definition_if_ctor 	 [Unstable]

get_type 	 [Unstable]

get_type_name_if_impl 	 [Unstable]

is_const_fn 	 [Unstable]

is_default_impl 	 [Unstable]

is_defaulted_trait 	 [Unstable]

is_impl 	 [Unstable]

is_staged_api 	 [Unstable]

is_static_method 	 [Unstable]

is_typedef 	 [Unstable]

maybe_get_item_ast 	 [Unstable]