Struct rustc_lint::metadata::cstore::crate_metadata

pub struct crate_metadata {
    pub name: String,
    pub local_path: RefCell<SmallVector<PathElem>>,
    pub local_def_path: RefCell<Vec<DisambiguatedDefPathData>>,
    pub data: MetadataBlob,
    pub cnum_map: RefCell<HashMap<u32, u32, DefaultState<FnvHasher>>>,
    pub cnum: u32,
    pub codemap_import_info: RefCell<Vec<ImportedFileMap>>,
    pub span: Span,
    pub staged_api: bool,
    pub index: Index,
    pub explicitly_linked: Cell<bool>,
}

Unstable (rustc_private)

Fields

name
Unstable (rustc_private)

local_path
Unstable (rustc_private)

local_def_path
Unstable (rustc_private)

data
Unstable (rustc_private)

cnum_map
Unstable (rustc_private)

cnum
Unstable (rustc_private)

codemap_import_info
Unstable (rustc_private)

span
Unstable (rustc_private)

staged_api
Unstable (rustc_private)

index
Unstable (rustc_private)

explicitly_linked
Unstable (rustc_private)

Flag if this crate is required by an rlib version of this crate, or in other words whether it was explicitly linked to. An example of a crate where this is false is when an allocator crate is injected into the dependency list, and therefore isn't actually needed to link an rlib.

Methods

impl crate_metadata

fn data(&'a self) -> &'a [u8]

Unstable (rustc_private)

fn name(&self) -> String

Unstable (rustc_private)

fn hash(&self) -> Svh

Unstable (rustc_private)

fn imported_filemaps(&'a self, codemap: &CodeMap) -> Ref<'a, Vec<ImportedFileMap>>

Unstable (rustc_private)

fn with_local_path<T, F>(&self, f: F) -> T where F: Fn(&[PathElem]) -> T

Unstable (rustc_private)

fn update_local_path(&self, candidate: Chain<Cloned<Iter<'a, PathElem>>, LinkedPath<'b>>)

Unstable (rustc_private)

fn local_def_path(&self) -> Vec<DisambiguatedDefPathData>

Unstable (rustc_private)

fn update_local_def_path(&self, candidate: Vec<DisambiguatedDefPathData>)

Unstable (rustc_private)

fn is_allocator(&self) -> bool

Unstable (rustc_private)

fn needs_allocator(&self) -> bool

Unstable (rustc_private)