Struct rustc_lint::metadata::index::Index

pub struct Index {
    // some fields omitted
}

Unstable (rustc_private)

The Item Index

This index maps the NodeId of each item to its location in the metadata.

The index is a sparse bit-vector consisting of a index-array and a position-array. Each entry in the index-array handles 32 nodes. The first word is a bit-array consisting of the nodes that hold items, the second is the index of the first of the items in the position-array. If there is a large set of non-item trailing nodes, they can be omitted from the index-array.

The index is serialized as an array of big-endian 32-bit words. The first word is the number of items in the position-array. Then, for each item, its position in the metadata follows. After that the index-array is stored.

struct index { u32 item_count; u32 items[self.item_count]; struct { u32 bits; u32 offset; } positions[..]; }

Methods

impl Index

fn lookup_item(&self, buf: &[u8], node: u32) -> Option<u32>

Unstable (rustc_private)

fn from_buf(buf: &[u8], start: usize, end: usize) -> Index

Unstable (rustc_private)