Module rustc_lint::middle

Unstable (rustc_private)

Modules

astconv_util 	 [Unstable]
This module contains a simple utility routine used by both typeck and const_eval. Almost certainly this could (and should) be refactored out of existence.

const_qualif 	 [Unstable]

const_val 	 [Unstable]

cstore 	 [Unstable]

dataflow 	 [Unstable]
A module for propagating forward dataflow information. The analysis assumes that the items to be propagated can be represented as bits and thus uses bitvectors. Your job is simply to specify the so-called GEN and KILL bits for each expression.

dead 	 [Unstable]

dependency_format 	 [Unstable]
Resolution of mixing rlibs and dylibs

effect 	 [Unstable]
Enforces the Rust effect system. Currently there is just one effect, unsafe.

entry 	 [Unstable]

expr_use_visitor 	 [Unstable]
A different sort of visitor for walking fn bodies. Unlike the normal visitor, which just walks the entire body in one shot, the ExprUseVisitor determines how expressions are being used.

free_region 	 [Unstable]
This file handles the relationships between free regions -- meaning lifetime parameters. Ordinarily, free regions are unrelated to one another, but they can be related via implied or explicit bounds. In that case, we track the bounds using the TransitiveRelation type and use that to decide when one free region outlives another and so forth.

intrinsicck 	 [Unstable]

lang_items 	 [Unstable]

liveness 	 [Unstable]
A classic liveness analysis based on dataflow over the AST. Computes, for each local variable in a function, whether that variable is live at a given point. Program execution points are identified by their id.

mem_categorization 	 [Unstable]

privacy 	 [Unstable]
A pass that checks to make sure private fields and methods aren't used outside their scopes. This pass will also generate a set of exported items which are available for use externally when compiled as a library.

reachable 	 [Unstable]

recursion_limit 	 [Unstable]

region 	 [Unstable]
This file actually contains two passes related to regions. The first pass builds up the scope_map, which describes the parent links in the region hierarchy. The second pass infers which types must be region parameterized.

resolve_lifetime 	 [Unstable]
Name resolution for lifetimes.

stability 	 [Unstable]
A pass that annotates every item and method with its stability level, propagating default levels lexically from parent to children ast nodes.

weak_lang_items 	 [Unstable]
Validity checking for weak lang items