Module rustc_lint::middle::check_match

Unstable (rustc_private)

Structs

MatchCheckCtxt 	 [Unstable]

StaticInliner 	 [Unstable]

Enums

Constructor 	 [Unstable]

Constants

DUMMY_WILD_PAT 	 [Unstable]

Functions

check_crate 	 [Unstable]

constructor_arity 	 [Unstable]
This computes the arity of a constructor. The arity of a constructor is how many subpattern patterns of that constructor should be expanded to.

specialize 	 [Unstable]
This is the main specialization step. It expands the first pattern in the given row into arity patterns based on the constructor. For most patterns, the step is trivial, for instance tuple patterns are flattened and box patterns expand into their inner pattern.