Enum rustc_lint::middle::const_eval::EvalHint

pub enum EvalHint<'tcx> {
    ExprTypeChecked,
    UncheckedExprHint(&'tcx TyS<'tcx>),
    UncheckedExprNoHint,
}

Unstable (rustc_private)

Hint to determine how to evaluate constant expressions which might not be type-checked.

Variants

ExprTypeChecked
Unstable (rustc_private)

We have a type-checked expression.

UncheckedExprHint
Unstable (rustc_private)

We have an expression which hasn't been type-checked, but we have an idea of what the type will be because of the context. For example, the length of an array is always usize. (This is referred to as a hint because it isn't guaranteed to be consistent with what type-checking would compute.)

UncheckedExprNoHint
Unstable (rustc_private)

We have an expression which has not yet been type-checked, and and we have no clue what the type will be.