Module rustc_lint::middle::cstore

Unstable (rustc_private)

Modules

tls 	 [Unstable]
Metadata encoding and decoding can make use of thread-local encoding and decoding contexts. These allow implementers of serialize::Encodable and Decodable to access information and datastructures that would otherwise not be available to them. For example, we can automatically translate def-id and span information during decoding because the decoding context knows which crate the data is decoded from. Or it allows to make ty::Ty decodable because the context has access to the TyCtxt that is needed for creating ty::Ty instances.

Structs

ChildItem 	 [Unstable]

CrateSource 	 [Unstable]

DummyCrateStore 	 [Unstable]
A dummy crate store that does not support any non-local crates, for test purposes.

ExternCrate 	 [Unstable]

LinkMeta 	 [Unstable]

Enums

DefLike 	 [Unstable]

FoundAst 	 [Unstable]

InlinedItem 	 [Unstable]
The data we save and restore about an inlined item or method. This is not part of the AST that we parse from a file, but it becomes part of the tree that we trans.

InlinedItemRef 	 [Unstable]
A borrowed version of hir::InlinedItem.

LinkagePreference 	 [Unstable]

NativeLibraryKind 	 [Unstable]

Constants

LOCAL_CRATE 	 [Unstable]
Item definitions in the currently-compiled crate would have the CrateNum LOCAL_CRATE in their DefId.

Traits

CrateStore 	 [Unstable]
A store of Rust crates, through with their metadata can be accessed.

Functions

validate_crate_name 	 [Unstable]