Trait rustc_lint::middle::cstore::CrateStore

pub trait CrateStore<'tcx>: Any {
    fn stability(&self, def: DefId) -> Option<Stability>;
    fn deprecation(&self, def: DefId) -> Option<Deprecation>;
    fn closure_kind(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> ClosureKind;
    fn closure_ty(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> ClosureTy<'tcx>;
    fn item_variances(&self, def: DefId) -> ItemVariances;
    fn repr_attrs(&self, def: DefId) -> Vec<ReprAttr>;
    fn item_type(&self, tcx: &ctxt<'tcx>, def: DefId) -> TypeScheme<'tcx>;
    fn item_path(&self, def: DefId) -> Vec<PathElem>;
    fn extern_item_path(&self, def: DefId) -> Vec<PathElem>;
    fn item_name(&self, def: DefId) -> Name;
    fn item_predicates(&self, tcx: &ctxt<'tcx>, def: DefId) -> GenericPredicates<'tcx>;
    fn item_super_predicates(&self, tcx: &ctxt<'tcx>, def: DefId) -> GenericPredicates<'tcx>;
    fn item_attrs(&self, def_id: DefId) -> Vec<Spanned<Attribute_>>;
    fn item_symbol(&self, def: DefId) -> String;
    fn trait_def(&self, tcx: &ctxt<'tcx>, def: DefId) -> TraitDef<'tcx>;
    fn adt_def(&self, tcx: &ctxt<'tcx>, def: DefId) -> &'tcx AdtDefData<'tcx, 'tcx>;
    fn method_arg_names(&self, did: DefId) -> Vec<String>;
    fn inherent_implementations_for_type(&self, def_id: DefId) -> Vec<DefId>;
    fn implementations_of_trait(&self, def_id: DefId) -> Vec<DefId>;
    fn provided_trait_methods(&self, tcx: &ctxt<'tcx>, def: DefId) -> Vec<Rc<Method<'tcx>>>;
    fn trait_item_def_ids(&self, def: DefId) -> Vec<ImplOrTraitItemId>;
    fn impl_items(&self, impl_def_id: DefId) -> Vec<ImplOrTraitItemId>;
    fn impl_trait_ref(&self, tcx: &ctxt<'tcx>, def: DefId) -> Option<TraitRef<'tcx>>;
    fn impl_polarity(&self, def: DefId) -> Option<ImplPolarity>;
    fn custom_coerce_unsized_kind(&self, def: DefId) -> Option<CustomCoerceUnsized>;
    fn associated_consts(&self, tcx: &ctxt<'tcx>, def: DefId) -> Vec<Rc<AssociatedConst<'tcx>>>;
    fn trait_of_item(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> Option<DefId>;
    fn impl_or_trait_item(&self, tcx: &ctxt<'tcx>, def: DefId) -> ImplOrTraitItem<'tcx>;
    fn is_const_fn(&self, did: DefId) -> bool;
    fn is_defaulted_trait(&self, did: DefId) -> bool;
    fn is_impl(&self, did: DefId) -> bool;
    fn is_default_impl(&self, impl_did: DefId) -> bool;
    fn is_extern_fn(&self, tcx: &ctxt<'tcx>, did: DefId) -> bool;
    fn is_static(&self, did: DefId) -> bool;
    fn is_static_method(&self, did: DefId) -> bool;
    fn is_statically_included_foreign_item(&self, id: u32) -> bool;
    fn is_typedef(&self, did: DefId) -> bool;
    fn dylib_dependency_formats(&self, cnum: u32) -> Vec<(u32, LinkagePreference)>;
    fn lang_items(&self, cnum: u32) -> Vec<(DefIndex, usize)>;
    fn missing_lang_items(&self, cnum: u32) -> Vec<LangItem>;
    fn is_staged_api(&self, cnum: u32) -> bool;
    fn is_explicitly_linked(&self, cnum: u32) -> bool;
    fn is_allocator(&self, cnum: u32) -> bool;
    fn crate_attrs(&self, cnum: u32) -> Vec<Spanned<Attribute_>>;
    fn crate_name(&self, cnum: u32) -> String;
    fn crate_hash(&self, cnum: u32) -> Svh;
    fn crate_struct_field_attrs(&self, cnum: u32) -> HashMap<DefId, Vec<Spanned<Attribute_>>, DefaultState<FnvHasher>>;
    fn plugin_registrar_fn(&self, cnum: u32) -> Option<DefId>;
    fn native_libraries(&self, cnum: u32) -> Vec<(NativeLibraryKind, String)>;
    fn reachable_ids(&self, cnum: u32) -> Vec<DefId>;
    fn def_path(&self, def: DefId) -> Vec<DisambiguatedDefPathData>;
    fn tuple_struct_definition_if_ctor(&self, did: DefId) -> Option<DefId>;
    fn struct_field_names(&self, def: DefId) -> Vec<Name>;
    fn item_children(&self, did: DefId) -> Vec<ChildItem>;
    fn crate_top_level_items(&self, cnum: u32) -> Vec<ChildItem>;
    fn maybe_get_item_ast(&'tcx self, tcx: &ctxt<'tcx>, def: DefId) -> FoundAst<'tcx>;
    fn maybe_get_item_mir(&self, tcx: &ctxt<'tcx>, def: DefId) -> Option<Mir<'tcx>>;
    fn crates(&self) -> Vec<u32>;
    fn used_libraries(&self) -> Vec<(String, NativeLibraryKind)>;
    fn used_link_args(&self) -> Vec<String>;
    fn metadata_filename(&self) -> &str;
    fn metadata_section_name(&self, target: &Target) -> &str;
    fn encode_type(&self, tcx: &ctxt<'tcx>, ty: &'tcx TyS<'tcx>) -> Vec<u8>;
    fn used_crates(&self, prefer: LinkagePreference) -> Vec<(u32, Option<PathBuf>)>;
    fn used_crate_source(&self, cnum: u32) -> CrateSource;
    fn extern_mod_stmt_cnum(&self, emod_id: u32) -> Option<u32>;
    fn encode_metadata(&self, tcx: &ctxt<'tcx>, reexports: &HashMap<u32, Vec<Export>, DefaultState<FnvHasher>>, item_symbols: &RefCell<HashMap<u32, String, DefaultState<FnvHasher>>>, link_meta: &LinkMeta, reachable: &HashSet<u32, DefaultState<FnvHasher>>, mir_map: &HashMap<u32, Mir<'tcx>, DefaultState<FnvHasher>>, krate: &Crate) -> Vec<u8>;
    fn metadata_encoding_version(&self) -> &[u8];
}

Unstable (rustc_private)

A store of Rust crates, through with their metadata can be accessed.

The : Any bound is a temporary measure that allows access to the backing rustc_metadata::cstore::CStore object. It will be removed in the near future - if you need to access internal APIs, please tell us.

Required Methods

fn stability(&self, def: DefId) -> Option<Stability>

Unstable (rustc_private)

fn deprecation(&self, def: DefId) -> Option<Deprecation>

Unstable (rustc_private)

fn closure_kind(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> ClosureKind

Unstable (rustc_private)

fn closure_ty(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> ClosureTy<'tcx>

Unstable (rustc_private)

fn item_variances(&self, def: DefId) -> ItemVariances

Unstable (rustc_private)

fn repr_attrs(&self, def: DefId) -> Vec<ReprAttr>

Unstable (rustc_private)

fn item_type(&self, tcx: &ctxt<'tcx>, def: DefId) -> TypeScheme<'tcx>

Unstable (rustc_private)

fn item_path(&self, def: DefId) -> Vec<PathElem>

Unstable (rustc_private)

fn extern_item_path(&self, def: DefId) -> Vec<PathElem>

Unstable (rustc_private)

fn item_name(&self, def: DefId) -> Name

Unstable (rustc_private)

fn item_predicates(&self, tcx: &ctxt<'tcx>, def: DefId) -> GenericPredicates<'tcx>

Unstable (rustc_private)

fn item_super_predicates(&self, tcx: &ctxt<'tcx>, def: DefId) -> GenericPredicates<'tcx>

Unstable (rustc_private)

fn item_attrs(&self, def_id: DefId) -> Vec<Spanned<Attribute_>>

Unstable (rustc_private)

fn item_symbol(&self, def: DefId) -> String

Unstable (rustc_private)

fn trait_def(&self, tcx: &ctxt<'tcx>, def: DefId) -> TraitDef<'tcx>

Unstable (rustc_private)

fn adt_def(&self, tcx: &ctxt<'tcx>, def: DefId) -> &'tcx AdtDefData<'tcx, 'tcx>

Unstable (rustc_private)

fn method_arg_names(&self, did: DefId) -> Vec<String>

Unstable (rustc_private)

fn inherent_implementations_for_type(&self, def_id: DefId) -> Vec<DefId>

Unstable (rustc_private)

fn implementations_of_trait(&self, def_id: DefId) -> Vec<DefId>

Unstable (rustc_private)

fn provided_trait_methods(&self, tcx: &ctxt<'tcx>, def: DefId) -> Vec<Rc<Method<'tcx>>>

Unstable (rustc_private)

fn trait_item_def_ids(&self, def: DefId) -> Vec<ImplOrTraitItemId>

Unstable (rustc_private)

fn impl_items(&self, impl_def_id: DefId) -> Vec<ImplOrTraitItemId>

Unstable (rustc_private)

fn impl_trait_ref(&self, tcx: &ctxt<'tcx>, def: DefId) -> Option<TraitRef<'tcx>>

Unstable (rustc_private)

fn impl_polarity(&self, def: DefId) -> Option<ImplPolarity>

Unstable (rustc_private)

fn custom_coerce_unsized_kind(&self, def: DefId) -> Option<CustomCoerceUnsized>

Unstable (rustc_private)

fn associated_consts(&self, tcx: &ctxt<'tcx>, def: DefId) -> Vec<Rc<AssociatedConst<'tcx>>>

Unstable (rustc_private)

fn trait_of_item(&self, tcx: &ctxt<'tcx>, def_id: DefId) -> Option<DefId>

Unstable (rustc_private)

fn impl_or_trait_item(&self, tcx: &ctxt<'tcx>, def: DefId) -> ImplOrTraitItem<'tcx>

Unstable (rustc_private)

fn is_const_fn(&self, did: DefId) -> bool

Unstable (rustc_private)

fn is_defaulted_trait(&self, did: DefId) -> bool

Unstable (rustc_private)

fn is_impl(&self, did: DefId) -> bool

Unstable (rustc_private)

fn is_default_impl(&self, impl_did: DefId) -> bool

Unstable (rustc_private)

fn is_extern_fn(&self, tcx: &ctxt<'tcx>, did: DefId) -> bool

Unstable (rustc_private)

fn is_static(&self, did: DefId) -> bool

Unstable (rustc_private)

fn is_static_method(&self, did: DefId) -> bool

Unstable (rustc_private)

fn is_statically_included_foreign_item(&self, id: u32) -> bool

Unstable (rustc_private)

fn is_typedef(&self, did: DefId) -> bool

Unstable (rustc_private)

fn dylib_dependency_formats(&self, cnum: u32) -> Vec<(u32, LinkagePreference)>

Unstable (rustc_private)

fn lang_items(&self, cnum: u32) -> Vec<(DefIndex, usize)>

Unstable (rustc_private)

fn missing_lang_items(&self, cnum: u32) -> Vec<LangItem>

Unstable (rustc_private)

fn is_staged_api(&self, cnum: u32) -> bool

Unstable (rustc_private)

fn is_explicitly_linked(&self, cnum: u32) -> bool

Unstable (rustc_private)

fn is_allocator(&self, cnum: u32) -> bool

Unstable (rustc_private)

fn crate_attrs(&self, cnum: u32) -> Vec<Spanned<Attribute_>>

Unstable (rustc_private)

fn crate_name(&self, cnum: u32) -> String

Unstable (rustc_private)

fn crate_hash(&self, cnum: u32) -> Svh

Unstable (rustc_private)

fn crate_struct_field_attrs(&self, cnum: u32) -> HashMap<DefId, Vec<Spanned<Attribute_>>, DefaultState<FnvHasher>>

Unstable (rustc_private)

fn plugin_registrar_fn(&self, cnum: u32) -> Option<DefId>

Unstable (rustc_private)

fn native_libraries(&self, cnum: u32) -> Vec<(NativeLibraryKind, String)>

Unstable (rustc_private)

fn reachable_ids(&self, cnum: u32) -> Vec<DefId>

Unstable (rustc_private)

fn def_path(&self, def: DefId) -> Vec<DisambiguatedDefPathData>

Unstable (rustc_private)

fn tuple_struct_definition_if_ctor(&self, did: DefId) -> Option<DefId>

Unstable (rustc_private)

fn struct_field_names(&self, def: DefId) -> Vec<Name>

Unstable (rustc_private)

fn item_children(&self, did: DefId) -> Vec<ChildItem>

Unstable (rustc_private)

fn crate_top_level_items(&self, cnum: u32) -> Vec<ChildItem>

Unstable (rustc_private)

fn maybe_get_item_ast(&'tcx self, tcx: &ctxt<'tcx>, def: DefId) -> FoundAst<'tcx>

Unstable (rustc_private)

fn maybe_get_item_mir(&self, tcx: &ctxt<'tcx>, def: DefId) -> Option<Mir<'tcx>>

Unstable (rustc_private)

fn crates(&self) -> Vec<u32>

Unstable (rustc_private)

fn used_libraries(&self) -> Vec<(String, NativeLibraryKind)>

Unstable (rustc_private)

fn used_link_args(&self) -> Vec<String>

Unstable (rustc_private)

fn metadata_filename(&self) -> &str

Unstable (rustc_private)

fn metadata_section_name(&self, target: &Target) -> &str

Unstable (rustc_private)

fn encode_type(&self, tcx: &ctxt<'tcx>, ty: &'tcx TyS<'tcx>) -> Vec<u8>

Unstable (rustc_private)

fn used_crates(&self, prefer: LinkagePreference) -> Vec<(u32, Option<PathBuf>)>

Unstable (rustc_private)

fn used_crate_source(&self, cnum: u32) -> CrateSource

Unstable (rustc_private)

fn extern_mod_stmt_cnum(&self, emod_id: u32) -> Option<u32>

Unstable (rustc_private)

fn encode_metadata(&self, tcx: &ctxt<'tcx>, reexports: &HashMap<u32, Vec<Export>, DefaultState<FnvHasher>>, item_symbols: &RefCell<HashMap<u32, String, DefaultState<FnvHasher>>>, link_meta: &LinkMeta, reachable: &HashSet<u32, DefaultState<FnvHasher>>, mir_map: &HashMap<u32, Mir<'tcx>, DefaultState<FnvHasher>>, krate: &Crate) -> Vec<u8>

Unstable (rustc_private)

fn metadata_encoding_version(&self) -> &[u8]

Unstable (rustc_private)