Enum rustc_lint::middle::def::Def

pub enum Def {
    DefFn(DefId, bool),
    DefSelfTy(Option<DefId>, Option<(u32, u32)>),
    DefMod(DefId),
    DefForeignMod(DefId),
    DefStatic(DefId, bool),
    DefConst(DefId),
    DefAssociatedConst(DefId),
    DefLocal(DefId, u32),
    DefVariant(DefId, DefId, bool),
    DefTy(DefId, bool),
    DefAssociatedTy(DefId, DefId),
    DefTrait(DefId),
    DefPrimTy(PrimTy),
    DefTyParam(ParamSpace, u32, DefId, Name),
    DefUpvar(DefId, u32, usize, u32),
    DefStruct(DefId),
    DefLabel(u32),
    DefMethod(DefId),
    DefErr,
}

Unstable (rustc_private)

Variants

DefFn
Unstable (rustc_private)

DefSelfTy
Unstable (rustc_private)

DefMod
Unstable (rustc_private)

DefForeignMod
Unstable (rustc_private)

DefStatic
Unstable (rustc_private)

DefConst
Unstable (rustc_private)

DefAssociatedConst
Unstable (rustc_private)

DefLocal
Unstable (rustc_private)

DefVariant
Unstable (rustc_private)

DefTy
Unstable (rustc_private)

DefAssociatedTy
Unstable (rustc_private)

DefTrait
Unstable (rustc_private)

DefPrimTy
Unstable (rustc_private)

DefTyParam
Unstable (rustc_private)

DefUpvar
Unstable (rustc_private)

DefStruct
Unstable (rustc_private)

Note that if it's a tuple struct's definition, the node id of the DefId may either refer to the item definition's id or the VariantData.ctor_id.

The cases that I have encountered so far are (this is not exhaustive): - If it's a ty_path referring to some tuple struct, then DefMap maps it to a def whose id is the item definition's id. - If it's an ExprPath referring to some tuple struct, then DefMap maps it to a def whose id is the VariantData.ctor_id.

DefLabel
Unstable (rustc_private)

DefMethod
Unstable (rustc_private)

DefErr
Unstable (rustc_private)

Methods

impl Def

fn var_id(&self) -> u32

Unstable (rustc_private)

fn def_id(&self) -> DefId

Unstable (rustc_private)

fn variant_def_ids(&self) -> Option<(DefId, DefId)>

Unstable (rustc_private)