Function rustc_lint::middle::infer::drain_fulfillment_cx

pub fn drain_fulfillment_cx<'tcx, T>(infcx: &InferCtxt<'a, 'tcx>, fulfill_cx: &mut FulfillmentContext<'tcx>, result: &T) -> Result<T, Vec<FulfillmentError<'tcx>>> where T: TypeFoldable<'tcx>

Unstable

Finishes processes any obligations that remain in the fulfillment context, and then "freshens" and returns result. This is primarily used during normalization and other cases where processing the obligations in fulfill_cx may cause type inference variables that appear in result to be unified, and hence we need to process those obligations to get the complete picture of the type.