Module rustc_lint::middle::resolve_lifetime

Unstable (rustc_private)

Name resolution for lifetimes.

Name resolution for lifetimes follows MUCH simpler rules than the full resolve. For example, lifetime names are never exported or used between functions, and they operate in a purely top-down way. Therefore we break lifetime name resolution into a separate pass.

Enums

DefRegion 	 [Unstable]

Functions

early_bound_lifetimes 	 [Unstable]

krate 	 [Unstable]

Type Definitions

NamedRegionMap 	 [Unstable]