Enum rustc_lint::middle::traits::Vtable

pub enum Vtable<'tcx, N> {
    VtableImpl(VtableImplData<'tcx, N>),
    VtableDefaultImpl(VtableDefaultImplData<N>),
    VtableParam(Vec<N>),
    VtableObject(VtableObjectData<'tcx>),
    VtableBuiltin(VtableBuiltinData<N>),
    VtableClosure(VtableClosureData<'tcx, N>),
    VtableFnPointer(&'tcx TyS<'tcx>),
}

Unstable (rustc_private)

Given the successful resolution of an obligation, the Vtable indicates where the vtable comes from. Note that while we call this a "vtable", it does not necessarily indicate dynamic dispatch at runtime. Vtable instances just tell the compiler where to find methods, but in generic code those methods are typically statically dispatched -- only when an object is constructed is a Vtable instance reified into an actual vtable.

For example, the vtable may be tied to a specific impl (case A), or it may be relative to some bound that is in scope (case B).

impl<T:Clone> Clone<T> for Option<T> { ... } // Impl_1
impl<T:Clone> Clone<T> for Box<T> { ... }    // Impl_2
impl Clone for int { ... }             // Impl_3

fn foo<T:Clone>(concrete: Option<Box<int>>,
                param: T,
                mixed: Option<T>) {

   // Case A: Vtable points at a specific impl. Only possible when
   // type is concretely known. If the impl itself has bounded
   // type parameters, Vtable will carry resolutions for those as well:
   concrete.clone(); // Vtable(Impl_1, [Vtable(Impl_2, [Vtable(Impl_3)])])

   // Case B: Vtable must be provided by caller. This applies when
   // type is a type parameter.
   param.clone();    // VtableParam

   // Case C: A mix of cases A and B.
   mixed.clone();    // Vtable(Impl_1, [VtableParam])
}

The type parameter N

See explanation on VtableImplData.

Variants

VtableImpl
Unstable (rustc_private)

Vtable identifying a particular impl.

VtableDefaultImpl
Unstable (rustc_private)

Vtable for default trait implementations This carries the information and nested obligations with regards to a default implementation for a trait Trait. The nested obligations ensure the trait implementation holds for all the constituent types.

VtableParam
Unstable (rustc_private)

Successful resolution to an obligation provided by the caller for some type parameter. The Vec<N> represents the obligations incurred from normalizing the where-clause (if any).

VtableObject
Unstable (rustc_private)

Virtual calls through an object

VtableBuiltin
Unstable (rustc_private)

Successful resolution for a builtin trait.

VtableClosure
Unstable (rustc_private)

Vtable automatically generated for a closure. The def ID is the ID of the closure expression. This is a VtableImpl in spirit, but the impl is generated by the compiler and does not appear in the source.

VtableFnPointer
Unstable (rustc_private)

Same as above, but for a fn pointer type with the given signature.

Methods

impl<'tcx, N> Vtable<'tcx, N>

fn nested_obligations(self) -> Vec<N>

Unstable (rustc_private)

fn map<M, F>(self, f: F) -> Vtable<'tcx, M> where F: FnMut(N) -> M

Unstable (rustc_private)