Struct rustc_lint::middle::traits::specialization_graph::Ancestors

pub struct Ancestors<'a, 'tcx> where 'tcx: 'a {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl<'a, 'tcx> Ancestors<'a, 'tcx>

fn type_defs(self, tcx: &'a TyCtxt<'tcx>, name: Name) -> TypeDefs<'a, 'tcx>

Unstable (rustc_private)

Search the items from the given ancestors, returning each type definition with the given name.

fn fn_defs(self, tcx: &'a TyCtxt<'tcx>, name: Name) -> FnDefs<'a, 'tcx>

Unstable (rustc_private)

Search the items from the given ancestors, returning each fn definition with the given name.

fn const_defs(self, tcx: &'a TyCtxt<'tcx>, name: Name) -> ConstDefs<'a, 'tcx>

Unstable (rustc_private)

Search the items from the given ancestors, returning each const definition with the given name.