Struct rustc_lint::middle::traits::SelectionContext

pub struct SelectionContext<'cx, 'tcx> where 'tcx: 'cx {
    // some fields omitted
}

Unstable

Methods

impl<'cx, 'tcx> SelectionContext<'cx, 'tcx>

fn new(infcx: &'cx InferCtxt<'cx, 'tcx>) -> SelectionContext<'cx, 'tcx>

Unstable

fn intercrate(infcx: &'cx InferCtxt<'cx, 'tcx>) -> SelectionContext<'cx, 'tcx>

Unstable

fn infcx(&self) -> &'cx InferCtxt<'cx, 'tcx>

Unstable

fn tcx(&self) -> &'cx ctxt<'tcx>

Unstable

fn param_env(&self) -> &'cx ParameterEnvironment<'cx, 'tcx>

Unstable

fn closure_typer(&self) -> &'cx InferCtxt<'cx, 'tcx>

Unstable

fn select(&mut self, obligation: &Obligation<'tcx, Binder<TraitPredicate<'tcx>>>) -> Result<Option<Vtable<'tcx, Obligation<'tcx, Predicate<'tcx>>>>, SelectionError<'tcx>>

Unstable

Attempts to satisfy the obligation. If successful, this will affect the surrounding type environment by performing unification.

fn evaluate_obligation(&mut self, obligation: &Obligation<'tcx, Predicate<'tcx>>) -> bool

Unstable

Evaluates whether the obligation obligation can be satisfied (by any means).

fn evaluate_impl(&mut self, impl_def_id: DefId, obligation: &Obligation<'tcx, Binder<TraitPredicate<'tcx>>>) -> bool

Unstable

Evaluates whether the impl with id impl_def_id could be applied to the self type obligation_self_ty. This can be used either for trait or inherent impls.