Struct rustc_lint::middle::ty::FnSig

pub struct FnSig<'tcx> {
    pub inputs: Vec<&'tcx TyS<'tcx>>,
    pub output: FnOutput<'tcx>,
    pub variadic: bool,
}

Unstable

Signature of a function type, which I have arbitrarily decided to use to refer to the input/output types.

inputs is the list of arguments and their modes.

output is the return type.

variadic indicates whether this is a variadic function. (only true for foreign fns)

Fields

inputs
Unstable

output
Unstable

variadic
Unstable