Struct rustc_lint::middle::ty::ItemSubsts

pub struct ItemSubsts<'tcx> {
    pub substs: Substs<'tcx>,
}

Unstable

Records the substitutions used to translate the polytype for an item into the monotype of an item reference.

Fields

substs
Unstable

Methods

impl<'tcx> ItemSubsts<'tcx>

fn empty() -> ItemSubsts<'tcx>

Unstable

fn is_noop(&self) -> bool

Unstable