Struct rustc_lint::middle::ty::TyIVar

pub struct TyIVar<'tcx, 'lt> where 'lt: 'tcx(_, _);

Unstable (rustc_private)

An IVar that contains a Ty. 'lt is a (reverse-variant) upper bound on the lifetime of the IVar. This is required because of variance problems: the IVar needs to be variant with respect to 'tcx (so it can be referred to from Ty) but can only be modified if its lifetime is exactly 'tcx.

Safety invariants: (A) self.0, if fulfilled, is a valid Ty<'tcx> (B) no aliases to this value with a 'tcx longer than this value's 'lt exist

NonZero is used rather than Unique because Unique isn't Copy.

Methods

impl<'tcx, 'lt> TyIVar<'tcx, 'lt>

fn new() -> TyIVar<'tcx, 'lt>

Unstable (rustc_private)

fn get(&self) -> Option<&'tcx TyS<'tcx>>

Unstable (rustc_private)

fn unwrap(&self) -> &'tcx TyS<'tcx>

Unstable (rustc_private)

fn fulfill(&self, value: &'lt TyS<'lt>)

Unstable (rustc_private)