Struct rustc_lint::middle::ty::TypeFlags

pub struct TypeFlags {
    // some fields omitted
}

Unstable

Methods

impl TypeFlags

const HAS_PARAMS: TypeFlags =

Unstable

const HAS_SELF: TypeFlags =

Unstable

const HAS_TY_INFER: TypeFlags =

Unstable

const HAS_RE_INFER: TypeFlags =

Unstable

const HAS_RE_EARLY_BOUND: TypeFlags =

Unstable

const HAS_FREE_REGIONS: TypeFlags =

Unstable

const HAS_TY_ERR: TypeFlags =

Unstable

const HAS_PROJECTION: TypeFlags =

Unstable

const HAS_TY_CLOSURE: TypeFlags =

Unstable

const HAS_LOCAL_NAMES: TypeFlags =

Unstable

const NEEDS_SUBST: TypeFlags =

Unstable

const NOMINAL_FLAGS: TypeFlags =

Unstable

const SIZEDNESS_CACHED: TypeFlags =

Unstable

const IS_SIZED: TypeFlags =

Unstable

const MOVENESS_CACHED: TypeFlags =

Unstable

const MOVES_BY_DEFAULT: TypeFlags =

Unstable

fn empty() -> TypeFlags

Unstable

Returns an empty set of flags.

fn all() -> TypeFlags

Unstable

Returns the set containing all flags.

fn bits(&self) -> u32

Unstable

Returns the raw value of the flags currently stored.

fn from_bits(bits: u32) -> Option<TypeFlags>

Unstable

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u32) -> TypeFlags

Unstable

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Unstable

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Unstable

Returns true if all flags are currently set.

fn intersects(&self, other: TypeFlags) -> bool

Unstable

Returns true if there are flags common to both self and other.

fn contains(&self, other: TypeFlags) -> bool

Unstable

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: TypeFlags)

Unstable

Inserts the specified flags in-place.

fn remove(&mut self, other: TypeFlags)

Unstable

Removes the specified flags in-place.

fn toggle(&mut self, other: TypeFlags)

Unstable

Toggles the specified flags in-place.