Function rustc_lint::middle::wf::trait_obligations

pub fn trait_obligations(infcx: &InferCtxt<'a, 'tcx>, body_id: u32, trait_ref: &TraitRef<'tcx>, span: Span, rfc1214: bool) -> Vec<Obligation<'tcx, Predicate<'tcx>>>

Unstable (rustc_private)

Returns the obligations that make this trait reference well-formed. For example, if there is a trait Set defined like trait Set<K:Eq>, then the trait reference Foo: Set<Bar> is WF if Bar: Eq.