rustc_lint
Module noop_method_call

Structs

NoopMethodCall

Statics

NOOP_METHOD_CALL

The noop_method_call lint detects specific calls to noop methods such as a calling <&T as Clone>::clone where T: !Clone.

SUSPICIOUS_DOUBLE_REF_OP

The suspicious_double_ref_op lint checks for usage of .clone()/.borrow()/.deref() on an &&T when T: !Deref/Borrow/Clone, which means the call will return the inner &T, instead of performing the operation on the underlying T and can be confusing.