Module rustc_lint::opaque_hidden_inferred_bound

Structs

AddBound 🔒

OpaqueHiddenInferredBound

OpaqueHiddenInferredBoundLint 🔒

Constants

_DERIVE_rustc_errors_AddToDiagnostic_FOR_AddBound 🔒

_DERIVE_rustc_errors_DecorateLint_a_FOR_OpaqueHiddenInferredBoundLint 🔒

Statics

OPAQUE_HIDDEN_INFERRED_BOUND

The opaque_hidden_inferred_bound lint detects cases in which nested impl Trait in associated type bounds are not written generally enough to satisfy the bounds of the associated type.