rustc_lint

Module static_mut_refs

Structs

StaticMutRefs

Statics

STATIC_MUT_REFS

The static_mut_refs lint checks for shared or mutable references of mutable static inside unsafe blocks and unsafe functions.

Functions

emit_static_mut_refs 🔒

path_is_static_mut 🔒