rustc_lint

Module transmute

Structs

CheckTransmutes 🔒

UndefinedTransmuteLint 🔒

Statics

PTR_TO_INTEGER_TRANSMUTE_IN_CONSTS

The ptr_to_integer_transmute_in_consts lint detects pointer to integer transmute in const functions and associated constants.

UNNECESSARY_TRANSMUTES

The unnecessary_transmutes lint detects transmutations that have safer alternatives.

Functions

check_ptr_transmute_in_const 🔒

Check for transmutes that exhibit undefined behavior. For example, transmuting pointers to integers in a const context.

check_unnecessary_transmute 🔒

Check for transmutes that overlap with stdlib methods. For example, transmuting [u8; 4] to u32.