Struct rustc_lint::types::ImproperCTypesVisitor

struct ImproperCTypesVisitor<'a, 'tcx> {
    cx: &'a LateContext<'tcx>,
    mode: CItemKind,
}

Fields
cx: &'a LateContext<'tcx>mode: CItemKind
Implementations

impl<'a, 'tcx> ImproperCTypesVisitor<'a, 'tcx>

fn check_for_array_ty(&mut self, sp: Span, ty: Ty<'tcx>) -> bool

Check if the type is array and emit an unsafe type lint.

fn check_field_type_for_ffi( &self, cache: &mut FxHashSet<Ty<'tcx>>, field: &FieldDef, args: GenericArgsRef<'tcx> ) -> FfiResult<'tcx>

Checks if the given field’s type is “ffi-safe”.

fn check_variant_for_ffi( &self, cache: &mut FxHashSet<Ty<'tcx>>, ty: Ty<'tcx>, def: AdtDef<'tcx>, variant: &VariantDef, args: GenericArgsRef<'tcx> ) -> FfiResult<'tcx>

Checks if the given VariantDef’s field types are “ffi-safe”.

fn check_type_for_ffi( &self, cache: &mut FxHashSet<Ty<'tcx>>, ty: Ty<'tcx> ) -> FfiResult<'tcx>

Checks if the given type is “ffi-safe” (has a stable, well-defined representation which can be exported to C code).

fn emit_ffi_unsafe_type_lint( &mut self, ty: Ty<'tcx>, sp: Span, note: DiagnosticMessage, help: Option<DiagnosticMessage> )

fn check_for_opaque_ty(&mut self, sp: Span, ty: Ty<'tcx>) -> bool

fn check_type_for_ffi_and_report_errors( &mut self, sp: Span, ty: Ty<'tcx>, is_static: bool, is_return_type: bool )

fn check_fn(&mut self, def_id: LocalDefId, decl: &'tcx FnDecl<'_>)

Check if a function’s argument types and result type are “ffi-safe”.

For a external ABI function, argument types and the result type are walked to find fn-ptr types that have external ABIs, as these still need checked.

fn check_foreign_fn(&mut self, def_id: LocalDefId, decl: &'tcx FnDecl<'_>)

Check if a function’s argument types and result type are “ffi-safe”.

fn check_foreign_static(&mut self, id: OwnerId, span: Span)

fn is_internal_abi(&self, abi: SpecAbi) -> bool

fn find_fn_ptr_ty_with_external_abi( &self, hir_ty: &Ty<'tcx>, ty: Ty<'tcx> ) -> Vec<(Ty<'tcx>, Span)>

Find any fn-ptr types with external ABIs in ty.

For example, Option<extern "C" fn()> returns extern "C" fn()