Module rustc_lint::unused

Structs

PathStatements

UnusedAllocation

UnusedBraces

UnusedImportBraces

UnusedParens

UnusedResults

Enums

UnusedDelimsCtx 🔒

Statics

PATH_STATEMENTS

The path_statements lint detects path statements with no effect.

UNUSED_ALLOCATION 🔒

The unused_allocation lint detects unnecessary allocations that can be eliminated.

UNUSED_BRACES 🔒

The unused_braces lint detects unnecessary braces around an expression.

UNUSED_IMPORT_BRACES 🔒

The unused_import_braces lint catches unnecessary braces around an imported item.

UNUSED_MUST_USE

The unused_must_use lint detects unused result of a type flagged as #[must_use].

UNUSED_PARENS 🔒

The unused_parens lint detects if, match, while and return with parentheses; they do not need them.

UNUSED_RESULTS

The unused_results lint checks for the unused result of an expression in a statement.

Traits

UnusedDelimLint 🔒

Used by both UnusedParens and UnusedBraces to prevent code duplication.