Trait rustc_lint::util::common::MemoizationMap

pub trait MemoizationMap {
    type Key: Clone;
    type Value: Clone;
    fn memoize<OP>(&self, key: Self::Key, op: OP) -> Self::Value where OP: FnOnce() -> Self::Value;
}

Unstable (rustc_private)

Associated Types

type Key: Clone

Unstable (rustc_private)

type Value: Clone

Unstable (rustc_private)

Required Methods

fn memoize<OP>(&self, key: Self::Key, op: OP) -> Self::Value where OP: FnOnce() -> Self::Value

Unstable (rustc_private)

If key is present in the map, return the valuee, otherwise invoke op and store the value in the map.

NB: if the receiver is a DepTrackingMap, special care is needed in the op to ensure that the correct edges are added into the dep graph. See the DepTrackingMap impl for more details!