Struct rustc_lint::util::nodemap::FnvHasher

pub struct FnvHasher(_);

Unstable

A speedy hash algorithm for node ids and def ids. The hashmap in libcollections by default uses SipHash which isn't quite as speedy as we want. In the compiler we're not really worried about DOS attempts, so we just default to a non-cryptographic hash.

This uses FNV hashing, as described here: http://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function