rustc_lint
Module dangling

Structs

DanglingPointerSearcher 🔒

This produces a dangling pointer:

DanglingPointers 🔒

FIXME: false negatives (i.e. the lint is not emitted when it should be)

Statics

DANGLING_POINTERS_FROM_TEMPORARIES

The dangling_pointers_from_temporaries lint detects getting a pointer to data of a temporary that will immediately get dropped.

Functions

is_temporary_rvalue 🔒

lint_expr 🔒

owns_allocation 🔒