rustc_lint
Module early

Implementation of the early lint pass.

The early lint pass works on AST nodes after macro expansion and name resolution, just before AST lowering. These lints are for purely syntactical lints.

Macros

impl_early_lint_pass 🔒

lint_callback 🔒

Structs

EarlyContextAndPass

Implements the AST traversal for early lint passes. T provides the check_* methods.

RuntimeCombinedEarlyLintPass 🔒

Traits

EarlyCheckNode

Early lints work on different nodes - either on the crate root, or on freshly loaded modules. This trait generalizes over those nodes.

Functions

check_ast_node

check_ast_node_inner 🔒