Module rustc_lint::levels

Structs

BuilderPush 🔒

LintLevelQueryMap 🔒

LintLevelSets 🔒

Collection of lint levels for the whole crate. This is used by AST-based lints, which do not wait until we have built HIR to be emitted.

LintLevelsBuilder

LintSet 🔒

Specifications found at this position in the stack. This map only represents the lints found for one set of attributes (like shallow_lint_levels_on does).

LintStackIndex 🔒

QueryMapExpectationsWrapper 🔒

TopDown

Constants

COMMAND_LINE 🔒

Traits

LintLevelsProvider

Functions

lint_expectations 🔒

parse_lint_and_tool_name

provide 🔒

shallow_lint_levels_on 🔒