rustc_lint
Module nonstandard_style

Structs

NonCamelCaseTypes

NonSnakeCase

NonUpperCaseGlobals

Enums

MethodLateContext 🔒

Statics

NON_CAMEL_CASE_TYPES

The non_camel_case_types lint detects types, variants, traits and type parameters that don’t have camel case names.

NON_SNAKE_CASE

The non_snake_case lint detects variables, methods, functions, lifetime parameters and modules that don’t have snake case names.

NON_UPPER_CASE_GLOBALS

The non_upper_case_globals lint detects static items that don’t have uppercase identifiers.

Functions

assoc_item_in_trait_impl 🔒

char_has_case 🔒

Some unicode characters have case, are considered upper case or lower case, but they can’t be upper cased or lower cased. For the purposes of the lint suggestion, we care about being able to change the char’s case.

is_camel_case 🔒

method_context 🔒

to_camel_case 🔒