Struct rustc_lint::LintStore

pub struct LintStore {
    lints: Vec<&'static Lint>,
    pub pre_expansion_passes: Vec<Box<dyn Fn() -> Box<dyn EarlyLintPass + 'static> + DynSend + DynSync>>,
    pub early_passes: Vec<Box<dyn Fn() -> Box<dyn EarlyLintPass + 'static> + DynSend + DynSync>>,
    pub late_passes: Vec<Box<dyn for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + DynSend + DynSync>>,
    pub late_module_passes: Vec<Box<dyn for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + DynSend + DynSync>>,
    by_name: FxHashMap<String, TargetLint>,
    lint_groups: FxHashMap<&'static str, LintGroup>,
}

Information about the registered lints.

This is basically the subset of Context that we can build early in the compile pipeline.

Fields
lints: Vec<&'static Lint>

Registered lints.

pre_expansion_passes: Vec<Box<dyn Fn() -> Box<dyn EarlyLintPass + 'static> + DynSend + DynSync>>

Constructor functions for each variety of lint pass.

These should only be called once, but since we want to avoid locks or interior mutability, we don’t enforce this (and lints should, in theory, be compatible with being constructed more than once, though not necessarily in a sane manner. This is safe though.)

early_passes: Vec<Box<dyn Fn() -> Box<dyn EarlyLintPass + 'static> + DynSend + DynSync>>late_passes: Vec<Box<dyn for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + DynSend + DynSync>>late_module_passes: Vec<Box<dyn for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + DynSend + DynSync>>

This is unique in that we construct them per-module, so not once.

by_name: FxHashMap<String, TargetLint>

Lints indexed by name.

lint_groups: FxHashMap<&'static str, LintGroup>

Map of registered lint groups to what lints they expand to.

Implementations

impl LintStore

pub fn new() -> LintStore

pub fn get_lints<'t>(&'t self) -> &'t [&'static Lint]

pub fn get_lint_groups<'t>( &'t self ) -> impl Iterator<Item = (&'static str, Vec<LintId>, bool)> + 't

pub fn register_early_pass( &mut self, pass: impl Fn() -> Box<dyn EarlyLintPass + 'static> + 'static + DynSend + DynSync )

pub fn register_pre_expansion_pass( &mut self, pass: impl Fn() -> Box<dyn EarlyLintPass + 'static> + 'static + DynSend + DynSync )

This lint pass is softly deprecated. It misses expanded code and has caused a few errors in the past. Currently, it is only used in Clippy. New implementations should avoid using this interface, as it might be removed in the future.

See rust#69838

See rust-clippy#5518

pub fn register_late_pass( &mut self, pass: impl for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + 'static + DynSend + DynSync )

pub fn register_late_mod_pass( &mut self, pass: impl for<'tcx> Fn(TyCtxt<'tcx>) -> Box<dyn LateLintPass<'tcx> + 'tcx> + 'static + DynSend + DynSync )

pub fn register_lints(&mut self, lints: &[&'static Lint])

Helper method for register_early/late_pass

pub fn register_group_alias( &mut self, lint_name: &'static str, alias: &'static str )

pub fn register_group( &mut self, from_plugin: bool, name: &'static str, deprecated_name: Option<&'static str>, to: Vec<LintId> )

pub fn register_ignored(&mut self, name: &str)

This lint should give no warning and have no effect.

This is used by rustc to avoid warning about old rustdoc lints before rustdoc registers them as tool lints.

pub fn register_renamed(&mut self, old_name: &str, new_name: &str)

This lint has been renamed; warn about using the new name and apply the lint.

pub fn register_removed(&mut self, name: &str, reason: &str)

pub fn find_lints(&self, lint_name: &str) -> Result<Vec<LintId>, FindLintError>

pub fn check_lint_name_cmdline( &self, sess: &Session, lint_name: &str, level: Level, registered_tools: &RegisteredTools )

Checks the validity of lint names derived from the command line.

pub fn is_lint_group(&self, lint_name: Symbol) -> bool

True if this symbol represents a lint group name.

pub fn check_lint_name( &self, lint_name: &str, tool_name: Option<Symbol>, registered_tools: &RegisteredTools ) -> CheckLintNameResult<'_>

Checks the name of a lint for its existence, and whether it was renamed or removed. Generates a DiagnosticBuilder containing a warning for renamed and removed lints. This is over both lint names from attributes and those passed on the command line. Since it emits non-fatal warnings and there are two lint passes that inspect attributes, this is only run from the late pass to avoid printing duplicate warnings.

fn no_lint_suggestion(&self, lint_name: &str) -> CheckLintNameResult<'_>

fn check_tool_name_for_backwards_compat( &self, lint_name: &str, tool_name: &str ) -> CheckLintNameResult<'_>